Moving linechart's zoom into internal subdirectory.
