+++ widgets/linechart/internal/zoom/zoom.go
