+++ widgets/linechart/internal/zoom/zoom_test.go
