+++ widgets/linechart/linechart.go
@@ -31,7 +31,7 @@ import (
 	"github.com/mum4k/termdash/internal/terminalapi"
 	"github.com/mum4k/termdash/internal/widgetapi"
 	"github.com/mum4k/termdash/widgets/linechart/internal/axes"
-	"github.com/mum4k/termdash/widgets/linechart/zoom"
+	"github.com/mum4k/termdash/widgets/linechart/internal/zoom"
 )
 
 // seriesValues represent values stored in the series.
