+++ widgets/linechart/options.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/mum4k/termdash/internal/cell"
 	"github.com/mum4k/termdash/widgets/linechart/internal/axes"
-	"github.com/mum4k/termdash/widgets/linechart/zoom"
+	"github.com/mum4k/termdash/widgets/linechart/internal/zoom"
 )
 
 // options.go contains configurable options for LineChart.
